requirements.txt
beautifulsoup4
ipykernel
ipython
ipywidgets
matplotlib
numpy
pandas
plotly-express
python-dotenv
requests
scikit-learn
scipy
seaborn
statsmodels
streamlit
torch
tqdm
yellowbrick